+++ src/main/kotlin/org/pcsoft/framework/kunit/distance/KLengthUnitBareValues.kt
@@ -67,3 +67,33 @@ val lightYears: KLengthUnitInstance = lengthOf(KDistanceUnit.LIGHT_YEAR.baseValu
 
 /** 1 parsec ([KDistanceUnit.PARSEC]). */
 val parsecs: KLengthUnitInstance = lengthOf(KDistanceUnit.PARSEC.baseValue)
+
+/** 1 cubit / Elle ([KDistanceUnit.CUBIT]). */
+val cubits: KLengthUnitInstance = lengthOf(KDistanceUnit.CUBIT.baseValue)
+
+/** 1 Roman foot / pes ([KDistanceUnit.ROMAN_FOOT]). */
+val romanFeet: KLengthUnitInstance = lengthOf(KDistanceUnit.ROMAN_FOOT.baseValue)
+
+/** 1 Roman pace / passus ([KDistanceUnit.ROMAN_PACE]). */
+val romanPaces: KLengthUnitInstance = lengthOf(KDistanceUnit.ROMAN_PACE.baseValue)
+
+/** 1 stadium ([KDistanceUnit.STADIUM]). */
+val stadia: KLengthUnitInstance = lengthOf(KDistanceUnit.STADIUM.baseValue)
+
+/** 1 Roman mile / mille passus ([KDistanceUnit.ROMAN_MILE]). */
+val romanMiles: KLengthUnitInstance = lengthOf(KDistanceUnit.ROMAN_MILE.baseValue)
+
+/** 1 rod / perch ([KDistanceUnit.ROD]). */
+val rods: KLengthUnitInstance = lengthOf(KDistanceUnit.ROD.baseValue)
+
+/** 1 league ([KDistanceUnit.LEAGUE]). */
+val leagues: KLengthUnitInstance = lengthOf(KDistanceUnit.LEAGUE.baseValue)
+
+/** 1 cable length ([KDistanceUnit.CABLE_LENGTH]). */
+val cableLengths: KLengthUnitInstance = lengthOf(KDistanceUnit.CABLE_LENGTH.baseValue)
+
+/** 1 verst / Werst ([KDistanceUnit.VERST]). */
+val versts: KLengthUnitInstance = lengthOf(KDistanceUnit.VERST.baseValue)
+
+/** 1 Prussian mile ([KDistanceUnit.PRUSSIAN_MILE]). */
+val prussianMiles: KLengthUnitInstance = lengthOf(KDistanceUnit.PRUSSIAN_MILE.baseValue)